:recycle: Split tokenizer builder funcs to toolkit
